+++ packages/revolution/test/revolution-token/Basic.t.sol
@@ -469,9 +469,9 @@ contract TokenBasicTest is RevolutionTokenTestSuite {
     function parseJson(
         string memory _json
     ) internal returns (string memory name, string memory description, string memory image) {
-        uint returnValue;
+        uint256 returnValue;
         JsmnSolLib.Token[] memory tokens;
-        uint actualNum;
+        uint256 actualNum;
 
         // Number of tokens to be parsed in the JSON (could be estimated or exactly known)
         uint256 numTokens = 20; // Increase if necessary to accommodate all fields in the JSON
